+++ apps/fluent-tester/src/FluentTesterApp.tsx
@@ -4,7 +4,6 @@ import * as React from 'react';
 import { Platform } from 'react-native';
 
 import { useHorizontalSizeClass } from '@fluentui-react-native/experimental-appearance-additions';
-import type { Theme } from '@fluentui-react-native/framework';
 import { ThemeReference, ThemeProvider } from '@fluentui-react-native/theme';
 
 import type { FluentTesterProps } from './FluentTester';
@@ -25,10 +24,9 @@ export const FluentTesterApp: React.FunctionComponent<FluentTesterProps> = (prop
         colors: { brandBackground2: 'red' }, // Overrides the buttonBackground color token, all other colors are kept in-tact.
       };
     },
-    (theme: Theme) => {
+    () => {
       return {
         colors: {
-          neutralForeground1: theme.colors.brandBackground2, // neutralBackground1 is 'red' in theme here because of previous recipe applied.
           hostColorPink: 'pink', // New custom color key.
           hostColorBrandText: 'purple', // New custom color key.
           hostColorButtonBackground: 'yellow', // New custom color key.