+++ packages/framework/framework/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log - @fluentui-react-native/framework
 
-This log was last generated on Tue, 07 Mar 2023 20:54:15 GMT and should not be manually modified.
+This log was last generated on Tue, 14 Mar 2023 20:50:45 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.9.6
+
+Tue, 14 Mar 2023 20:50:45 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.18.0
+- Bump @fluentui-react-native/theme-types to v0.31.0
+- Bump @fluentui-react-native/tokens to v0.20.12
 
 ## 0.9.5