+++ packages/theming/default-theme/CHANGELOG.md
@@ -1,8 +1,20 @@
 # Change Log - @fluentui-react-native/default-theme
 
-This log was last generated on Tue, 07 Mar 2023 20:54:15 GMT and should not be manually modified.
+This log was last generated on Tue, 14 Mar 2023 20:50:45 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.18.0
+
+Tue, 14 Mar 2023 20:50:45 GMT
+
+### Minor changes
+
+- Add numeric font family type (ruaraki@microsoft.com)
+- Bump @fluentui-react-native/theme-tokens to v0.24.5
+- Bump @fluentui-react-native/theme-types to v0.31.0
+- Bump @fluentui-react-native/theme to v0.8.3
+- Bump @fluentui-react-native/theming-utils to v0.23.8
 
 ## 0.17.5